+++ cnccode/controller_classes/CTCustomerReviewMeeting.inc.php
@@ -1279,7 +1279,7 @@ WHERE INTERNAL = 1 AND missing=0 AND os LIKE \'%server%\' and size >= 1024 AND c
         $firstTimeFixRequests = [
             "title"   => "Qualifying First Time Fix Requests",
             "columns" => ["Attempted", "Achieved"],
-            "data"    => [["Attempted", $data["attemptedFirstTimeFix"]], ["Achieved", $data["firstTimeFix"]]]
+            "data"    => [["Attempted - {$data["attemptedFirstTimeFix"]}", $data["attemptedFirstTimeFix"]], ["Achieved - {$data["firstTimeFix"]}", $data["firstTimeFix"]]]
         ];
         $sourceOfRequests     = [
             "title"   => "Source of Requests (%)",
@@ -1287,7 +1287,7 @@ WHERE INTERNAL = 1 AND missing=0 AND os LIKE \'%server%\' and size >= 1024 AND c
             "data"    => []
         ];
         foreach ($data["raiseTypeSummary"] as $item) {
-            $sourceOfRequests["data"] [] = [$item["description"], $item["total"]];
+            $sourceOfRequests["data"] [] = ["{$item["description"]} - {$item["total"]}", $item["total"]];
         }
         foreach ($historicData as $datum) {
             $row                       = [