+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project changes will be shown here.
 ## [unreleased]
 ### Added
 - Issue 1402 - Standard Rate Card.
+- Issue 1379 - Change status of Sales Orders & Purchase Orders
 ### Changed
 - Issue 1399 - Force Customer Name when creating Templated Sales Quotation.
 - Issue 1368 - convert StarterLeaverManagement.php,  RootCause.php, UtilityEmails.php, IgnoredADDomains.php, Office365Licenses.php and SRSource.php
@@ -150,7 +151,6 @@ This project changes will be shown here.
 ### Fixed
 - Issue 1226 - StreamOne error handling
 - Issue 1349 - Duplicate Duplicate Contact Warning.
-
 
 ## [v5.8.6] - 2021-04-22
 ### Fixed